requirements.txt
requests==2.26.0
Flask==3.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
Jinja2==3.1.3
openai==1.14.2
python-dotenv==1.0.1
rsa==4.9
wheel==0.43.0